Fixed issue where server stales if there is no config file
Also fixed local joining.
