+++ xlive/Globals.h
@@ -32,12 +32,5 @@ extern bool displayXyz;
 
 extern std::map<DWORD, bool> achievementList;
 
-//some utility functions below
-std::vector<std::string> &split(const std::string &s, char delim, std::vector<std::string> &elems);
-std::vector<std::string> split(const std::string &s, char delim);
-
 extern int H2GetInstanceId();
 
-int stripWhitespace(wchar_t *inputStr);
-void HexStrToBytes(const std::string& hexStr, BYTE* byteBuf, size_t bufLen);
-std::string ByteToHexStr(const BYTE* buffer, size_t size);
