+++ xlive/H2MOD/Modules/Startup/Startup.cpp
@@ -462,7 +462,15 @@ void InitH2Startup2() {
 		//currently credentials are taken from the config file.
 		//also don't enable this since nothing's initialised for the server.
 		AccountEdit_remember = false;
-		HandleGuiLogin(0, H2Config_login_identifier, H2Config_login_password, nullptr);
+		if (HandleGuiLogin(0, H2Config_login_identifier, H2Config_login_password, nullptr) == false)
+		{
+			// if login fails, setup default data, so the server will not stale
+			BYTE abEnet[6];
+			BYTE abOnline[20];
+			XNetRandom(abEnet, 6);
+			XNetRandom(abOnline, 20);
+			ConfigureUserDetails("[Username]", "12345678901234567890123456789012", rand(), 0, H2Config_ip_lan, ByteToHexStr(abEnet, 6).c_str(), ByteToHexStr(abOnline, 20).c_str(), false);
+		}
 	}
 }
